clean and fix tests, replace old snapshots
