+++ client/src/app/__tests__/PartitionList.test.jsx
@@ -1,0 +1,24 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+
+import PartitionList from '../components/PartitionList';
+
+describe('PartitionList unit tests', () => {
+  let wrapper;
+  const props = {
+    topicInfo: {
+      numberOfPartitions: 8,
+    },
+  };
+  beforeAll(() => {
+    wrapper = shallow(<PartitionList {...props} />);
+  });
+
+  it('Should render without errors', () => {
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('Should render a Partition component', () => {
+    expect(wrapper.exists('Partition')).toBe(true);
+  });
+});
